+++ CHANGES.rst
@@ -102,3 +102,4 @@ v1.0.11
 v1.0.12[unreleased]
 -----------
 * Add admin email
+* Add causes to Organization serializers